+++ services/ui-src/src/components/export/ExportedReportFieldTable.tsx
@@ -1,24 +1,25 @@
-import { ReactElement } from "react";
 // components
 import { Box, Heading } from "@chakra-ui/react";
-import { ExportedReportFieldRow, Table } from "components";
+import { Table } from "components";
 // types
 import {
   FormField,
   StandardReportPageShape,
   DrawerReportPageShape,
   FormLayoutElement,
-  isFieldElement,
   ReportType,
-  FieldChoice,
   ReportStatus,
   HeadingLevel,
   FormTableType,
 } from "types";
 // utils
-import { getReportVerbiage, useStore, parseCustomHtml } from "utils";
-// assets
-import { sxSharedExportStyles } from "components/pages/Export/ExportedReportPage";
+import { getReportVerbiage, useStore } from "utils";
+// helpers
+import {
+  sx,
+  renderFieldTableBody,
+  renderCalculationTables,
+} from "./exportFieldTableHelpers";
 
 export const ExportedReportFieldTable = ({
   section,
@@ -60,60 +61,6 @@ export const ExportedReportFieldTable = ({
       (table) => table.tableType === FormTableType.CALCULATION
     ) || [];
 
-  const renderCalculationTables = () => {
-    return calculationTables.map((table) => {
-      if (table.bodyRows) {
-        const bodyRows = renderServiceTableBody(table.bodyRows);
-        const footerRow = renderServiceTableBody(table.footRows);
-
-        let percentageValue;
-        if (section.name === "Qualified HCBS") {
-          percentageValue = fieldData?.["fmap_qualifiedHcbsPercentage"] || 100;
-        } else if (section.name === "Demonstration Services") {
-          percentageValue =
-            fieldData?.["fmap_demonstrationServicesPercentage"] || 100;
-        } else {
-          percentageValue = formPercentage;
-        }
-
-        const percentageText =
-          table.verbiage?.percentage || "[auto-populated]%";
-        const displayPercentage = percentageText.replace(
-          "{{percentage}}",
-          `${percentageValue}%`
-        );
-
-        return (
-          <Box key={table.id}>
-            <Heading as="h3" sx={sx.subHeading}>
-              {table.verbiage?.title}
-            </Heading>
-            {table.verbiage?.percentage && (
-              <Box sx={sx.tableSubHeading}>
-                {parseCustomHtml(displayPercentage)}
-              </Box>
-            )}
-            <Table
-              sx={{ ...sx.table, ...sx.serviceTable }}
-              content={{
-                headRow: [
-                  "Service",
-                  "Total Computable",
-                  "Total State / Territory Share",
-                  "Total Federal Share",
-                ],
-                bodyRows: bodyRows,
-                footRow: footerRow,
-              }}
-              data-testid={`service-table-${table.id}`}
-            />
-          </Box>
-        );
-      }
-      return null;
-    });
-  };
-
   // SAR "General Information" section layout is a unique case with multiple section headings within the same page
   if (reportType === ReportType.SAR && section.name === "General Information") {
     return renderGeneralInformation(
@@ -132,7 +79,8 @@ export const ExportedReportFieldTable = ({
 
   return (
     <>
-      {calculationTables.length > 0 && renderCalculationTables()}
+      {calculationTables.length > 0 &&
+        renderCalculationTables(section, fieldData, formPercentage)}
       {nonTableFields.length > 0 && (
         <>
           {nonTableFields?.[0]?.props?.title && (
@@ -239,130 +187,8 @@ export const renderGeneralInformation = (
   });
 };
 
-export const renderFieldTableBody = (
-  formFields: (FormField | FormLayoutElement)[],
-  pageType: string,
-  showHintText: boolean,
-  entityType?: string
-) => {
-  const tableRows: ReactElement[] = [];
-  // recursively renders field rows
-  const renderFieldRow = (
-    formField: FormField | FormLayoutElement,
-    parentFieldCheckedChoiceIds?: string[]
-  ) => {
-    tableRows.push(
-      <ExportedReportFieldRow
-        key={formField.id}
-        formField={formField}
-        pageType={pageType}
-        entityType={entityType}
-        parentFieldCheckedChoiceIds={parentFieldCheckedChoiceIds}
-        showHintText={showHintText}
-      />
-    );
-    // for drawer pages, render nested child field if any entity has a checked parent choice
-    if (pageType === "drawer") {
-      return;
-    }
-
-    // Handle rendering nested children
-    if (isFieldElement(formField) && formField.props?.choices) {
-      formField.props.choices.forEach((choice: FieldChoice) => {
-        if (choice.children) {
-          choice.children.forEach((c) => renderFieldRow(c));
-        }
-      });
-    }
-  };
-
-  // map through form fields and call renderer
-  formFields.map((field: FormField | FormLayoutElement) => {
-    if (isFieldElement(field)) {
-      renderFieldRow(field);
-    }
-  });
-  return tableRows;
-};
-
-export const renderServiceTableBody = (bodyRows: any) => {
-  const { report } = useStore();
-  return bodyRows.map((row: any) => {
-    const label = row[0];
-    const totalComputableField = row[1];
-    const totalStateTerritoryShareField = row[2];
-    const totalFederalShareField = row[3];
-
-    if (
-      totalComputableField?.id &&
-      totalStateTerritoryShareField?.id &&
-      totalFederalShareField?.id
-    ) {
-      const totalComputable =
-        report?.fieldData[totalComputableField.id] || "Not answered";
-      const totalStateTerritoryShare =
-        report?.fieldData[totalStateTerritoryShareField.id] || "$0";
-      const totalFederalShare =
-        report?.fieldData[totalFederalShareField.id] || "$0";
-
-      return [
-        label,
-        totalComputable,
-        totalStateTerritoryShare,
-        totalFederalShare,
-      ];
-    }
-    return [];
-  });
-};
-
 export interface Props {
   section: StandardReportPageShape | DrawerReportPageShape;
   showHintText?: boolean;
   headingLevel?: HeadingLevel;
 }
-
-const sx = {
-  table: sxSharedExportStyles.table,
-  serviceTable: {
-    "& th:not(:first-of-type), & td:not(:first-of-type)": {
-      textAlign: "right",
-    },
-    "& thead tr:first-of-type": {
-      borderBottom: "2px solid",
-      borderColor: "gray.400",
-    },
-    "& tbody tr td:nth-of-type(3), & tbody tr td:nth-of-type(4)": {
-      fontWeight: "bold",
-    },
-    "& tfoot tr td": {
-      fontWeight: "bold",
-      color: "black",
-    },
-    "& tfoot tr:first-of-type": {
-      borderTop: "2px solid",
-      borderColor: "gray.400",
-    },
-    "& tfoot tr:last-of-type": {
-      borderBottom: "2px solid",
-      borderColor: "gray.400",
-    },
-  },
-  heading: {
-    fontSize: "xl",
-    fontWeight: "bold",
-    color: "black",
-  },
-  subHeading: {
-    fontSize: "xl",
-    "& + table": {
-      marginTop: "spacer2",
-    },
-  },
-  tableSubHeading: {
-    fontSize: "md",
-    fontWeight: "bold",
-    color: "gray_dark",
-    marginTop: "spacer2",
-  },
-};
